requirements.txt
filetype==1.2.0
python-dotenv==1.0.1
python-gitlab==5.6.0
